Avoiding alloc/free in EM_step
The EM_step contains a for loop that basically implements the XD algorithm. This loop runs over all datapoints and, for each datapoint, allocates and free several vectors and matrices. In most use cases, the data size is constant over the datapoints, thus leading to unnecessary alloc and free which will slow down the code. I modified the code so that the alloc/free is done, for each thread, in a parallel block. Additionally, the code checks if the temporary vectors and matrices needs to be reallocated because of a change in the size of the datapoint.